Conflicts:
	include/fea/audio/openal.hpp
	src/audio/effectslot.cpp

+++ src/freetype-gl/texture-atlas.c
@@ -36,6 +36,7 @@
 #include <string.h>
 #include <assert.h>
 #include <limits.h>
+#include <stdint.h>
 #include <fea/rendering/opengl.hpp>
 #include <fea/freetype-gl/texture-atlas.h>
 
@@ -219,7 +220,7 @@ texture_atlas_get_region( texture_atlas_t * self,
 		if( y >= 0 )
 		{
             node = (ivec3 *) vector_get( self->nodes, i );
-			if( ( (y + height) < best_height ) ||
+			if( ( (int)(y + height) < best_height ) ||
                 ( ((y + height) == best_height) && (node->z < best_width)) )
 			{
 				best_height = y + height;
